part way through migrating auth util
